+++ psdaq/psdaq/control/control.py
@@ -16,9 +16,9 @@ import string
 from p4p.client.thread import Context
 from threading import Thread, Event, Condition
 import dgramCreate as dc
-
-PORT_BASE = 29980
-POSIX_TIME_AT_EPICS_EPOCH = 631152000
+from psdaq.control.DaqDefs import DaqDefs, create_msg, error_msg, warning_msg,  \
+                                  progress_msg, fileReport_msg, front_pub_port, \
+                                  back_pub_port, front_rep_port, back_pull_port
 
 report_keys = ['error', 'warning', 'fileReport']
 
@@ -231,400 +231,6 @@ class ConfigDBError(ControlError):
     """
     def __init__(self, message):
         self.message = message
-
-class DaqControl:
-    'Base class for controlling data acquisition'
-
-    # transitionId is a subset of the TransitionId.hh enum
-    transitionId = {
-        'ClearReadout'      : 0,
-        'Reset'             : 1,
-        'Configure'         : 2,
-        'Unconfigure'       : 3,
-        'BeginRun'          : 4,
-        'EndRun'            : 5,
-        'BeginStep'         : 6,
-        'EndStep'           : 7,
-        'Enable'            : 8,
-        'Disable'           : 9,
-        'SlowUpdate'        : 10,
-        'L1Accept'          : 12,
-    }
-
-    transitions = ['rollcall', 'alloc', 'dealloc',
-                   'connect', 'disconnect',
-                   'configure', 'unconfigure',
-                   'beginrun', 'endrun',
-                   'beginstep', 'endstep',
-                   'enable', 'disable',
-                   'slowupdate', 'reset']
-
-    states = [
-        'reset',
-        'unallocated',
-        'allocated',
-        'connected',
-        'configured',
-        'starting',
-        'paused',
-        'running'
-    ]
-
-    # default readout group is self.platform
-
-    def __init__(self, *, host, platform, timeout):
-        self.host = host
-        self.platform = platform
-        self.timeout = timeout
-
-        # initialize zmq socket
-        self.context = zmq.Context(1)
-        self.front_sub = self.context.socket(zmq.SUB)
-        self.front_sub.connect('tcp://%s:%d' % (host, front_pub_port(platform)))
-        self.front_sub.setsockopt(zmq.SUBSCRIBE, b'')
-        self.front_req = None
-        self.front_req_endpoint = 'tcp://%s:%d' % (host, front_rep_port(platform))
-        self.front_req_init()
-
-    #
-    # DaqControl.getState - get current state
-    #
-    def getState(self):
-        retval = 'error'
-        try:
-            msg = create_msg('getstate')
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except zmq.Again:
-            logging.error('getState() timeout (%.1f sec)' % (self.timeout / 1000.))
-            logging.info('getState() reinitializing zmq socket')
-            self.front_req_init()
-        except Exception as ex:
-            logging.error('getState() Exception: %s' % ex)
-        except KeyboardInterrupt:
-            print('KeyboardInterrupt')
-        else:
-            try:
-                retval = reply['header']['key']
-            except KeyError:
-                pass
-
-        return retval
-
-    #
-    # DaqControl.getPlatform - get platform
-    #
-    def getPlatform(self):
-        retval = {}
-        try:
-            msg = create_msg('getstate')
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except zmq.Again:
-            logging.error('getPlatform() timeout (%.1f sec)' % (self.timeout / 1000.))
-            logging.info('getPlatform() reinitializing zmq socket')
-            self.front_req_init()
-        except Exception as ex:
-            logging.error('getPlatform() Exception: %s' % ex)
-        except KeyboardInterrupt:
-            print('KeyboardInterrupt')
-        else:
-            try:
-                retval = reply['body']
-            except KeyError:
-                pass
-
-        return retval
-
-    #
-    # DaqControl.getJsonConfig - get json configuration
-    #
-    def getJsonConfig(self):
-        src = self.getPlatform()
-        dst = levels_to_activedet(src)
-        retval =  oldjson.dumps(dst, sort_keys=True, indent=4)
-        logging.debug('getJsonConfig() return value: %s' % retval)
-        return retval
-
-    #
-    # DaqControl.storeJsonConfig - store json configuration
-    #
-    def storeJsonConfig(self, json_data):
-        retval = {}
-        body = {"json_data": json_data}
-        try:
-            msg = create_msg('storejsonconfig', body=body)
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except zmq.Again:
-            logging.error('storeJsonConfig() timeout (%.1f sec)' % (self.timeout / 1000.))
-            logging.info('storeJsonConfig() reinitializing zmq socket')
-            self.front_req_init()
-        except Exception as ex:
-            logging.error('storeJsonConfig() Exception: %s' % ex)
-        except KeyboardInterrupt:
-            print('KeyboardInterrupt')
-        else:
-            try:
-                retval = reply['body']
-            except KeyError:
-                pass
-
-        return retval
-
-    #
-    # DaqControl.selectPlatform - select platform
-    #
-    def selectPlatform(self, body):
-        retval = {}
-        try:
-            msg = create_msg('selectplatform', body=body)
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except zmq.Again:
-            logging.error('selectPlatform() timeout (%.1f sec)' % (self.timeout / 1000.))
-            logging.info('selectPlatform() reinitializing zmq socket')
-            self.front_req_init()
-        except Exception as ex:
-            logging.error('selectPlatform() Exception: %s' % ex)
-        except KeyboardInterrupt:
-            print('KeyboardInterrupt')
-        else:
-            try:
-                retval = reply['body']
-            except KeyError:
-                pass
-
-        return retval
-
-    #
-    # DaqControl.getInstrument - get instrument name
-    #
-    def getInstrument(self):
-        r1 = None
-        try:
-            msg = create_msg('getinstrument')
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except Exception as ex:
-            print('getInstrument() Exception: %s' % ex)
-        else:
-            try:
-                r1 = reply['body']['instrument']
-            except Exception as ex:
-                print('getInstrument() Exception: %s' % ex)
-
-        return r1
-
-    #
-    # DaqControl.getStatus - get status
-    #
-    def getStatus(self):
-        r1 = r2 = r3 = r4 = r6 = 'error'
-        r5 = {}
-        r7 = r8 = r9 = 'error'
-        try:
-            msg = create_msg('getstatus')
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except Exception as ex:
-            print('getStatus() Exception: %s' % ex)
-        except KeyboardInterrupt:
-            print('KeyboardInterrupt')
-        else:
-            try:
-                r1 = reply['body']['transition']
-                r2 = reply['body']['state']
-                r3 = reply['body']['config_alias']
-                r4 = reply['body']['recording']
-                r5 = reply['body']['platform']
-                r6 = reply['body']['bypass_activedet']
-                r7 = reply['body']['experiment_name']
-                r8 = reply['body']['run_number']
-                r9 = reply['body']['last_run_number']
-            except KeyError:
-                pass
-
-        return (r1, r2, r3, r4, r5, r6, r7, r8, r9)
-
-    #
-    # DaqControl.monitorStatus - monitor the status
-    #
-    def monitorStatus(self):
-
-        # process messages
-        while True:
-            try:
-                msg = self.front_sub.recv_json()
-
-                if msg['header']['key'] == 'status':
-                    # return transition, state, config_alias, recording, bypass_activedet, experiment_name, run_number, last_run_number
-                    return msg['body']['transition'], msg['body']['state'], msg['body']['config_alias'], msg['body']['recording'], msg['body']['bypass_activedet'],\
-                           msg['body']['experiment_name'], msg['body']['run_number'], msg['body']['last_run_number']
-
-                elif msg['header']['key'] == 'error':
-                    # return 'error', error message, 'error', 'error', 'error', 'error', 'error', 'error'
-                    return 'error', msg['body']['err_info'], 'error', 'error', 'error', 'error', 'error', 'error'
-
-                elif msg['header']['key'] == 'warning':
-                    # return 'error', error message, 'error', 'error', 'error', 'error', 'error', 'error'
-                    return 'warning', msg['body']['err_info'], 'error', 'error', 'error', 'error', 'error', 'error'
-
-                elif msg['header']['key'] == 'fileReport':
-                    # return 'fileReport', path, 'error', 'error', 'error', 'error', 'error', 'error'
-                    return 'fileReport', msg['body']['path'], 'error', 'error', 'error', 'error', 'error', 'error'
-
-                elif msg['header']['key'] == 'progress':
-                    # return 'progress', transition, elapsed, total, 'error', 'error', 'error', 'error'
-                    return 'progress', msg['body']['transition'], msg['body']['elapsed'], msg['body']['total'], 'error', 'error', 'error', 'error'
-
-            except KeyboardInterrupt:
-                break
-
-            except KeyError as ex:
-                logging.error('KeyError: %s' % ex)
-                break
-
-        return None, None, None, None, None, None, None, None
-
-    #
-    # DaqControl.setState - change the state
-    # The optional second argument is a dictionary containing
-    # one entry per transition that contains information that
-    # will be put into the phase1-json of the transition. An example:
-    # {'beginstep': {'myvalue1':3 , 'myvalue2': {'myvalue3':72}},
-    #  'enable':    {'myvalue5':37, 'myvalue6': 'hello'}}
-    #
-    def setState(self, state, phase1Info={}):
-        errorMessage = None
-        try:
-            msg = create_msg('setstate.' + state, body=phase1Info)
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except zmq.Again:
-            errorMessage = 'setState() timeout (%.1f sec)' % (self.timeout / 1000.)
-            logging.info('setState() reinitializing zmq socket')
-            self.front_req_init()
-        except Exception as ex:
-            errorMessage = 'setState() Exception: %s' % ex
-        else:
-            try:
-                errorMessage = reply['body']['err_info']
-            except KeyError:
-                pass
-
-        return errorMessage
-
-    #
-    # DaqControl.setConfig - set BEAM/NOBEAM
-    #
-    def setConfig(self, config):
-        errorMessage = None
-        try:
-            msg = create_msg('setconfig.' + config)
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except Exception as ex:
-            errorMessage = 'setConfig() Exception: %s' % ex
-        else:
-            try:
-                errorMessage = reply['body']['err_info']
-            except KeyError:
-                pass
-
-        return errorMessage
-
-    #
-    # DaqControl.setRecord - set record flag
-    #   True or False
-    #
-    def setRecord(self, recordIn):
-        errorMessage = None
-        if type(recordIn) == type(True):
-            if recordIn:
-                record = '1'
-            else:
-                record = '0'
-
-            try:
-                msg = create_msg('setrecord.' + record)
-                self.front_req.send_json(msg)
-                reply = self.front_req.recv_json()
-            except Exception as ex:
-                errorMessage = 'setRecord() Exception: %s' % ex
-            else:
-                try:
-                    errorMessage = reply['body']['err_info']
-                except KeyError:
-                    pass
-        else:
-            errorMessage = 'setRecord() requires True or False'
-
-        return errorMessage
-
-    #
-    # DaqControl.setBypass - set bypass_activedet flag
-    #   True or False
-    #
-    def setBypass(self, bypassIn):
-        errorMessage = None
-        if type(bypassIn) == type(True):
-            if bypassIn:
-                bypass = '1'
-            else:
-                bypass = '0'
-
-            try:
-                msg = create_msg('setbypass.' + bypass)
-                self.front_req.send_json(msg)
-                reply = self.front_req.recv_json()
-            except Exception as ex:
-                errorMessage = 'setBypass() Exception: %s' % ex
-            else:
-                try:
-                    errorMessage = reply['body']['err_info']
-                except KeyError:
-                    pass
-        else:
-            errorMessage = 'setBypass() requires True or False'
-
-        return errorMessage
-
-    #
-    # DaqControl.setTransition - trigger a transition
-    # The optional second argument is a dictionary containing
-    # information that will be put into the phase1-json of the transition.
-    # An example:
-    # {'myvalue1':3 , 'myvalue2': {'myvalue3':72}}
-    #
-    def setTransition(self, transition, phase1Info={}):
-        errorMessage = None
-        try:
-            msg = create_msg(transition, body=phase1Info)
-            self.front_req.send_json(msg)
-            reply = self.front_req.recv_json()
-        except Exception as ex:
-            errorMessage = 'setTransition() Exception: %s' % ex
-        else:
-            try:
-                errorMessage = reply['body']['err_info']
-            except KeyError:
-                pass
-
-        return errorMessage
-
-    #
-    # DaqControl.front_req_init - (re)initialize the front_req zmq socket
-    #
-    def front_req_init(self):
-        # if socket previouly created, close it
-        if self.front_req is not None:
-            self.front_req.close()
-        # create new socket
-        self.front_req = self.context.socket(zmq.REQ)
-        self.front_req.linger = 0
-        self.front_req.RCVTIMEO = self.timeout
-        self.front_req.connect(self.front_req_endpoint)
 
 next_dict = {
     'reset' :       { 'unallocated' : 'rollcall',
@@ -934,44 +540,6 @@ def timestampStr():
     sec = int(current.timestamp()) - POSIX_TIME_AT_EPICS_EPOCH
     return '%010d-%09d' % (sec, nsec)
 
-def create_msg(key, msg_id=None, sender_id=None, body={}):
-    if msg_id is None:
-        msg_id = timestampStr()
-    msg = {'header': {
-               'key': key,
-               'msg_id': msg_id,
-               'sender_id': sender_id},
-           'body': body}
-    return msg
-
-def error_msg(message):
-    body = {'err_info': message}
-    return create_msg('error', body=body)
-
-def warning_msg(message):
-    body = {'err_info': message}
-    return create_msg('warning', body=body)
-
-def fileReport_msg(path):
-    body = {'path': path}
-    return create_msg('fileReport', body=body)
-
-def progress_msg(transition, elapsed, total):
-    body = {'transition': transition, 'elapsed': int(elapsed), 'total': int(total)}
-    return create_msg('progress', body=body)
-
-def back_pull_port(platform):
-    return PORT_BASE + platform
-
-def back_pub_port(platform):
-    return PORT_BASE + platform + 10
-
-def front_rep_port(platform):
-    return PORT_BASE + platform + 20
-
-def front_pub_port(platform):
-    return PORT_BASE + platform + 30
-
 def get_readout_group_mask(body):
     mask = 0
     if 'drp' in body:
@@ -1205,7 +773,7 @@ class CollectionManager():
         self.lastTransition = 'reset'
         self.recording = False
 
-        self.collectMachine = Machine(self, DaqControl.states, initial='reset')
+        self.collectMachine = Machine(self, DaqDefs.states, initial='reset')
 
         self.collectMachine.add_transition('reset', '*', 'reset',
                                            conditions='condition_reset', after='report_status')
@@ -1299,7 +867,7 @@ class CollectionManager():
                 else:
                     self.handle_setbypass(True)
                 answer = None
-            elif key[0] in DaqControl.transitions:
+            elif key[0] in DaqDefs.transitions:
                 # is body dict not-empty?
                 if body:
                     self.phase1Info[key[1]] = body
@@ -1470,7 +1038,7 @@ class CollectionManager():
         logging.debug('handle_setstate(\'%s\') in state %s' % (newstate, self.state))
         stateBefore = self.state
 
-        if newstate not in DaqControl.states:
+        if newstate not in DaqDefs.states:
             stateError = 'state \'%s\' not recognized' % newstate
             errMsg = stateError.replace("\"", "")
             logging.error(errMsg)
@@ -1734,12 +1302,12 @@ class CollectionManager():
         # ...clear readout
         self.pva.pv_put(self.pva.pvGroupL0Reset, self.groups)
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['ClearReadout'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['ClearReadout'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
         time.sleep(1.0)
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['BeginRun'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['BeginRun'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
 
@@ -1765,7 +1333,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['EndRun'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['EndRun'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
         self.step_groups(mask=0)    # default is no scanning
@@ -1792,7 +1360,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['BeginStep'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['BeginStep'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
 
@@ -1812,7 +1380,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['EndStep'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['EndStep'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
 
@@ -1829,7 +1397,7 @@ class CollectionManager():
         # phase 1 not needed
         # phase 2 no replies needed
         for pv in self.pva.pvListMsgHeader:
-            if not self.pva.pv_put(pv, DaqControl.transitionId['SlowUpdate']):
+            if not self.pva.pv_put(pv, DaqDefs.transitionId['SlowUpdate']):
                 update_ok = False
                 break
 
@@ -2402,13 +1970,13 @@ class CollectionManager():
         # ...clear readout
         self.pva.pv_put(self.pva.pvGroupL0Reset, self.groups)
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['ClearReadout'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['ClearReadout'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
         time.sleep(1.0)
         # ...configure
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['Configure'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['Configure'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
         self.step_groups(mask=0)    # default is no scanning
@@ -2434,7 +2002,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['Unconfigure'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['Unconfigure'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
 
@@ -2497,7 +2065,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['Enable'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['Enable'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
 
@@ -2537,7 +2105,7 @@ class CollectionManager():
 
         # phase 2
         for pv in self.pva.pvListMsgHeader:
-            self.pva.pv_put(pv, DaqControl.transitionId['Disable'])
+            self.pva.pv_put(pv, DaqDefs.transitionId['Disable'])
         self.pva.pv_put(self.pva.pvGroupMsgInsert, self.groups)
         self.pva.pv_put(self.pva.pvGroupMsgInsert, 0)
